NXP-20603: move s2 operation to automation features + move constants and helpers to SuggestConstants
