+++ nuxeo-distribution/nuxeo-distribution-server/pom.xml
@@ -356,10 +356,6 @@
       <artifactId>nuxeo-elasticsearch-seqgen</artifactId>
     </dependency>
     <dependency>
-      <groupId>org.nuxeo.ecm.automation</groupId>
-      <artifactId>nuxeo-automation-select2</artifactId>
-    </dependency>
-    <dependency>
       <groupId>org.nuxeo.ecm.platform</groupId>
       <artifactId>nuxeo-invite</artifactId>
     </dependency>
