--- nuxeo-features/nuxeo-automation/nuxeo-automation-select2/src/main/resources/META-INF/MANIFEST.MF
@@ -1,12 +1,0 @@
-Manifest-Version: 1.0
-Bundle-ManifestVersion: 2
-Bundle-Name: Nuxeo Select2 Automation
-Bundle-SymbolicName: org.nuxeo.ecm.automation.select2;singleton:=true
-Bundle-Vendor: Nuxeo
-Bundle-Localization: bundle
-Require-Bundle: org.nuxeo.ecm.automation.core
-Export-Package: org.nuxeo.ecm.automation.core
-Eclipse-LazyStart: true
-Bundle-ActivationPolicy: lazy
-Bundle-ClassPath: .
-Nuxeo-Component: OSGI-INF/operations-contrib.xml
