--- nuxeo-features/nuxeo-automation/nuxeo-automation-select2/src/main/resources/OSGI-INF/deployment-fragment.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0"?>
-<fragment version="1">
-
-</fragment>
