+++ pom.xml
@@ -1072,11 +1072,6 @@
       </dependency>
       <dependency>
         <groupId>org.nuxeo.ecm.automation</groupId>
-        <artifactId>nuxeo-automation-select2</artifactId>
-        <version>${nuxeo.features.version}</version>
-      </dependency>
-      <dependency>
-        <groupId>org.nuxeo.ecm.automation</groupId>
         <artifactId>nuxeo-automation-jsf</artifactId>
         <version>${nuxeo.features.version}</version>
       </dependency>
